extract Civil* conversion functions
renames some for clarity as well
