+++ packages/components/src/InputTime/InputTime.tsx
@@ -1,11 +1,55 @@
 import React, { useRef } from "react";
-import {
-  civilTimeToHTMLTime,
-  htmlTimeToCivilTime,
-} from "./civilTimeConversions";
-import { InputTimeProps } from "./InputTimeProps";
+import { CivilTime } from "@std-proposal/temporal";
 import { useTimePredict } from "./hooks/useTimePredict";
-import { FormField, FormFieldProps } from "../FormField";
+import {
+  civilTimeToTimeString,
+  timeStringToCivilTime,
+} from "../utils/civilTimeConversions";
+import { CommonFormFieldProps, FormField, FormFieldProps } from "../FormField";
+
+export interface InputTimeProps
+  extends Pick<
+      CommonFormFieldProps,
+      | "align"
+      | "description"
+      | "disabled"
+      | "invalid"
+      | "inline"
+      | "loading"
+      | "name"
+      | "onValidation"
+      | "placeholder"
+      | "size"
+      | "clearable"
+    >,
+    Pick<
+      FormFieldProps,
+      | "maxLength"
+      | "readonly"
+      | "autocomplete"
+      | "max"
+      | "min"
+      | "onEnter"
+      | "onFocus"
+      | "onBlur"
+      | "inputRef"
+      | "validations"
+    > {
+  /**
+   * Intial value of the input. Only use this when you need to prepopulate the
+   * field with a data that is not controlled by the components state. If a
+   * state is controlling the value, use the `value` prop instead.
+   */
+  readonly defaultValue?: CivilTime;
+  /**
+   * Set the component to the given value.
+   */
+  readonly value?: CivilTime;
+  /**
+   * Function called when user changes input value.
+   */
+  onChange?(newValue?: CivilTime): void;
+}
 
 export function InputTime({
   defaultValue,
@@ -18,8 +62,8 @@ export function InputTime({
 
   const fieldProps: FormFieldProps = {
     onChange: handleChange,
-    ...(defaultValue && { defaultValue: civilTimeToHTMLTime(defaultValue) }),
-    ...(!defaultValue && { value: civilTimeToHTMLTime(value) }),
+    ...(defaultValue && { defaultValue: civilTimeToTimeString(defaultValue) }),
+    ...(!defaultValue && { value: civilTimeToTimeString(value) }),
     ...params,
   };
 
@@ -37,7 +81,7 @@ export function InputTime({
   );
 
   function handleChange(newValue: string) {
-    onChange?.(htmlTimeToCivilTime(newValue));
+    onChange?.(timeStringToCivilTime(newValue));
   }
 
   function handleBlur() {
