+++ packages/components/src/InputTime/time-input-polyfill.d.ts
@@ -1,3 +1,4 @@
+// IS THIS POLYFILL NEEDED?
 // Declaring types for polyfill.
 // eslint-disable-next-line @typescript-eslint/no-extraneous-class
 declare class TimePolyfill {
